Extract tracker code into its own module
