+++ packages/tracker/package.json
@@ -1,0 +1,26 @@
+{
+  "name": "@umami/tracker",
+  "version": "3.0.3",
+  "description": "Umami analytics tracker core library",
+  "author": "Umami Software, Inc. <hello@umami.is>",
+  "license": "MIT",
+  "homepage": "https://umami.is",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/umami-software/umami.git",
+    "directory": "packages/tracker"
+  },
+  "type": "module",
+  "main": "./dist/index.js",
+  "types": "./dist/index.d.ts",
+  "exports": {
+    ".": {
+      "types": "./dist/index.d.ts",
+      "import": "./dist/index.js"
+    }
+  },
+  "files": [
+    "dist"
+  ],
+  "sideEffects": false
+}
