+++ packages/tracker/src/index.ts
@@ -1,0 +1,12 @@
+export { createTracker } from './tracker';
+
+export type {
+  CustomEventFunction,
+  EventData,
+  EventProperties,
+  PageViewProperties,
+  TrackedProperties,
+  TrackerConfig,
+  UmamiTracker,
+  WithRequired,
+} from './types';
