+++ rollup.tracker.config.js
@@ -1,6 +1,7 @@
 import 'dotenv/config';
 import replace from '@rollup/plugin-replace';
 import terser from '@rollup/plugin-terser';
+import typescript from '@rollup/plugin-typescript';
 
 export default {
   input: 'src/tracker/index.js',
@@ -9,6 +10,12 @@ export default {
     format: 'iife',
   },
   plugins: [
+    typescript({
+      tsconfig: 'packages/tracker/tsconfig.json',
+      declaration: false,
+      declarationMap: false,
+      sourceMap: false,
+    }),
     replace({
       __COLLECT_API_HOST__: process.env.COLLECT_API_HOST || '',
       __COLLECT_API_ENDPOINT__: process.env.COLLECT_API_ENDPOINT || '/api/send',
